fix: align token transfer calls and mock to u128 after integer overflow audit
PR #134 updated TokenInterface to use u128 but left stale i128 casts in
lib.rs transfer call sites and the inline MockToken in test.rs, breaking
compilation. Removes the as i128 casts from all three transfer calls and
migrates the unit-test mock (mint/transfer/balance) to u128 throughout so
the over-allocation tests in tests/over_allocation.rs can compile and run.

Closes #67

+++ contracts/wave_milestone/src/lib.rs
@@ -125,7 +125,7 @@ impl WaveMilestoneContract {
 
         // ── Fund transfer ──
         let token = TokenClient::new(&env, &asset);
-        token.transfer(&maintainer, &env.current_contract_address(), &(total_funds as i128));
+        token.transfer(&maintainer, &env.current_contract_address(), &total_funds);
 
         // ── Persist pool ──
         let pool = MilestonePool {
@@ -219,7 +219,7 @@ impl WaveMilestoneContract {
 
         // ── Transfer tokens ──
         let token = TokenClient::new(&env, &pool.asset);
-        token.transfer(&env.current_contract_address(), &developer, &(amount as i128));
+        token.transfer(&env.current_contract_address(), &developer, &amount);
 
         // ── Update pool state ──
         pool.allocated_funds = pool.allocated_funds.checked_add(amount).ok_or(Error::InvalidAmount)?;
@@ -280,7 +280,7 @@ impl WaveMilestoneContract {
         }
 
         let token = TokenClient::new(&env, &pool.asset);
-        token.transfer(&env.current_contract_address(), &maintainer, &(remaining as i128));
+        token.transfer(&env.current_contract_address(), &maintainer, &remaining);
 
         pool.total_funds = pool.allocated_funds;
         env.storage().instance().set(&DataKey::Pool, &pool);
